move the `input_handler` module from `nu-command` to `nu-cmd-base`
