+++ crates/nu-command/Cargo.toml
@@ -96,6 +96,7 @@ url = "2.2"
 uuid = { version = "1.3", features = ["v4"] }
 wax = { version = "0.5" }
 which = { version = "4.4", optional = true }
+nu-cmd-base = { version = "0.80.1", path = "../nu-cmd-base" }
 
 [target.'cfg(windows)'.dependencies]
 winreg = "0.50"
