+++ crates/nu-command/src/bytes/at.rs
@@ -1,7 +1,5 @@
-use crate::{
-    input_handler::{operate, CmdArgument},
-    util,
-};
+use crate::util;
+use nu_cmd_base::input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::{
     ast::{Call, CellPath},
