+++ crates/nu-command/src/conversions/into/int.rs
@@ -1,6 +1,6 @@
 use chrono::{FixedOffset, TimeZone};
 
-use crate::input_handler::{operate, CmdArgument};
+use nu_cmd_base::input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::{
     ast::{Call, CellPath},
