+++ crates/nu-command/src/strings/encode_decode/hex.rs
@@ -1,4 +1,4 @@
-use crate::input_handler::{operate as general_operate, CmdArgument};
+use nu_cmd_base::input_handler::{operate as general_operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::ast::{Call, CellPath};
 use nu_protocol::engine::{EngineState, Stack};
