+++ crates/nu-command/src/strings/str_/length.rs
@@ -1,5 +1,5 @@
 use crate::grapheme_flags;
-use crate::input_handler::{operate, CmdArgument};
+use nu_cmd_base::input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
 use nu_protocol::ast::CellPath;
