+++ crates/nu-command/src/strings/str_/replace.rs
@@ -1,5 +1,5 @@
-use crate::input_handler::{operate, CmdArgument};
 use fancy_regex::{NoExpand, Regex};
+use nu_cmd_base::input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::{
     ast::{Call, CellPath},
