+++ crates/nu-command/src/strings/str_/substring.rs
@@ -1,5 +1,5 @@
-use crate::input_handler::{operate, CmdArgument};
 use crate::{grapheme_flags, util};
+use nu_cmd_base::input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
 use nu_protocol::ast::CellPath;
